Minor test fixes
